implement a getter to value
