+++ docs/index.html
@@ -7,8 +7,8 @@
 
   <body>
     <mn-form>
-      <mn-checkbox name="gender" value="female" readonly></mn-checkbox>
-      <mn-checkbox name="gender" value="male" checked readonly></mn-checkbox>
+      <mn-checkbox name="house" value="stark"></mn-checkbox>
+      <mn-checkbox name="house" value="lannister" checked></mn-checkbox>
     </mn-form>
     <!-- <h2>vanilla js</h2>
     <mn-form name="form">
